refactor: move ProcessInstanceDependant to base package
