+++ operate/webapp/src/main/java/io/camunda/operate/webapp/opensearch/writer/OpensearchProcessInstanceWriter.java
@@ -10,7 +10,7 @@ package io.camunda.operate.webapp.opensearch.writer;
 import io.camunda.operate.conditions.OpensearchCondition;
 import io.camunda.operate.store.ProcessStore;
 import io.camunda.operate.webapp.elasticsearch.reader.ProcessInstanceReader;
-import io.camunda.webapps.schema.descriptors.operate.ProcessInstanceDependant;
+import io.camunda.webapps.schema.descriptors.ProcessInstanceDependant;
 import io.camunda.webapps.schema.descriptors.template.ListViewTemplate;
 import io.camunda.webapps.schema.descriptors.template.OperationTemplate;
 import io.camunda.webapps.schema.entities.listview.ProcessInstanceForListViewEntity;
